Bug#36158022: Replace the Robin Hood library [patch 1, SpillState, noclose]
Reduce the number of files that use the robin-hood-hashing library, so
that there are fewer places to modify when replacing it with the new
library.

The materialize_iterator::SpillState class is moved from
composite_iterators.h to an anonymous namespace in the corresponding
.cc file, since that's the only place it is used. This removes the
dependency on robin_hood.h from the header file, and all the files
that include composite_iterators.h directly or indirectly.

Also remove an unused reference to it from hash_join_iterator.cc.

Change-Id: I1f8051c8144966cb386c306fc37ef09479b3f912