Bug#36158022: Replace the Robin Hood library [patch 2, HashJoinRowBuffer, noclose]
Reduce the number of files that use the robin-hood-hashing library, so
that there are fewer places to modify when replacing it with the new
library.

This patch moves the definition of the hash map used by
HashJoinRowBuffer from hash_join_buffer.h to hash_join_buffer.cc, so
that there is no reference to robin_hood.h from the header file. This
reduced the number of files that depend on robin_hood.h directly or
indirectly.

Change-Id: I9544b8fac9747dd53883f229ac9bb456f4a52900
(cherry picked from commit ad0ac49b1827033cf2086b9b43d506a67315b92a)