WL#16005: Index-level hints for hypergraph optimizer [4, SKIP_SCAN]
Implementation of skip scan hints - SKIP_SCAN, NO_SKIP_SCAN
Implementation of skip_scan optimizer switch

SKIP_SCAN - Forces optimizer to pick a skip scan plan using
any of the indexes specified in the hint for that table
if possible. If indexes are not specified, optimizer is free
to choose the optimal index of its choice to do the skip scan.
If it's not possible to do the skip scan, it picks the next
best possible plan to access the table.
While finding index range scans in FindIndexRangeScans(),
if skip scan is forced, it first proposes skip scan plans.
If such a plan is not found, then it proposes range scans,
followed by index merge scans.
Index lookups, covering index scans and table scans are
proposed if no skip scan is found.

NO_SKIP_SCAN - Forces optimizer to not pick a skip scan plan
involving any of the indexes specified in the hint for that
table. If no indexes are specified, skip scan is not picked
for that table.
No additional changes needed to enforce this hint.

skip_scan optimizer switch enables/disables index skip scan
optimization. Note that this switch applies for index skip
scan and not skip scans used for handling group by.
We now check the switch and the hint before proposing a
skip scan access path.

Change-Id: Iaae37d250d0c2293ace7085886992f48bc442517