Bug#36155387: Hypergraph: Too many rows from index               range scan for JSON_CONTAINS
For the json function JSON_CONTAINS, range optimizer
constructs an inexact range. However, the tree is
not marked inexact. As a result, hypergraph optimizer
does not know that it needs to place the filter on
top of the range access for it to get the correct
result. Old optimizer always places the full filter
on top of any range access. So we get the correct
result with it.
E.g For a filter like this one -
   json_contains(j->'$[*]', '[1,2,3]');
Range optimizer constructs a range tree like this:
(3 MEMBER OF (json_extract(j,_utf8mb4'$[*]'))) OR
(2 MEMBER OF (json_extract(j,_utf8mb4'$[*]'))) OR
(1 MEMBER OF (json_extract(j,_utf8mb4'$[*]')))
So, at the end of the range access, we have rows
that have value [1] or [2] or [3] for the column
"j". However, we want rows which have all three
values for the column "j". This necessitates the
need for the final filter on top of the range
access even in hypergraph optimizer.
Solution is to mark range trees constructed using
multi valued keyparts as inexact.

Change-Id: I5cd808a58139eb1e9f53c7554537aec22f0f8dad