BUG#36267511 ndb_binlog_index test timeout on fast machines
Problem:
When running the ndb_binlog_index.test on fast machines it generates
progressively more binlog. The generated binlog is then analyzed for
correctness using `mysqlbinlog --verbose`, LOAD DATA INFILE and SELECT's
with several LIKE. On a fast machine the file size exceeds half a million
rows and thus takes significant time to analyze, causing test to
eventually time out.

Analysis:
Rewriting the way that the generated binlog is analyzed is not feasible
due to complexity of the correctness checks. Instead the number of rows
and work needed to analyze need to be kept constant.
Assuming machines will get faster in the future it's not viable to just
reduce test runtime as the same problem would soon manifest itself.
Instead the amount of work should be regulated by stopping test when
certain amount of binlog has been generated.

Solution:
Rewrite binlog generator to only run until binlog size exceeds a given
size or max time exceeded. The problematic size on fast machine was 20Mb
and thus a threshold of 2Mb binlog or at least 4 laps AND max 25
seconds (for slow machine), this should allow both slow and fast
machines to run the test succesfully.

Change-Id: I28cb7b9f161834ef583746212b02d91e138c812f