Bug#36158022: Replace the Robin Hood library [patch 3, tests, noclose]
Some tests that use EXCEPT or INTERSECT are non-deterministic because
they depend on a particular ordering of the rows even though there is
no ORDER BY clause. The ordering of the results changes when the hash
map implementation is changed. Make these tests more robust and ready
for using the new library.

Change-Id: I41a51328f5f059009a04fcbdb3537e0927ca4982