WL#16005: Index-level hints for hypergraph optimizer [6, INDEX/NO_INDEX]
Implement INDEX/NO_INDEX hints. This also implements
FORCE INDEX, USE INDEX and IGNORE INDEX.
Forces optimizer to use a particular index to
access the table for any scope. Since hypergraph
optimizer already looks into "keys_in_use_for_query"
for proposing index access methods, this index hint
is already implemented partially.
Change is needed in that we propose table scans
at the end instead of proposing them ahead of
index scans. This results in some plan changes
in tests where a covering index scan and table
scan have the same cost. A table scan is proposed
only when the forced index could not be used
in any way to access the table. A non-covering
index scan is not preferred over a table scan
even if the index is forced. This is inline with
the behaviour in the old optimizer.

Change-Id: I1e2e5cc7d5222d55109994e762dd8144132c4686