Bug#35483764 Binlogging stalls after filling large EventBuffer
The NdbApi EventBuffer internally pools free memory to
reduce the interactions with the runtime / OS, while allowing
memory that is no longer needed to be returned for other uses.

This amount of pooled free memory is subtracted from the total
allocated memory to determine an 'inuse memory' value which is
reported and used for e.g. enforcing buffer limits.

The amount of pooled free memory was represented using a 32bit
value, so that if it exceeded 4GB, the value wrapped, and the
free memory appeared to be reduced.

This affected :
 - EventBuffer memory release to the runtime/OS
 - EventBuffer free memory reporting
 - EventBuffer memory limit handling

This is fixed by using a 64bit value to represent the amount
of pooled free memory.

Change-Id: If304e08e82c70ff075d1466a1c6f256efa0c9996