Bug#36191153: Improve comment for FunctionalDependency:FD
Clarify a comment in interesting_orders.h that explains what a
functional dependency is. Also change "row stream" to "tuple stream" a
couple of places, since "tuple stream" is used more in this file, and
is what is used in two of the academic papers referenced in this file.

Also add a definition of "sort-ahead" in build_interesting_orders.h,
since that term is used a lot without a proper definition.

Change-Id: I4c94a762a973b51f8242d2914227cd15147babfb