Bug#113442 Fix innodb corruption that is caused when a column order alter is allowed to proceed with ALGORITHM=INSTANT
ALTER commands that change column order are not allowed to use
ALGORITHM=INSTANT:

    mysql> ALTER TABLE test.`t` MODIFY COLUMN c4 datetime after c1, ALGORITHM=INSTANT;
    ERROR 1845 (0A000): ALGORITHM=INSTANT is not supported for this operation. Try ALGORITHM=COPY/INPLACE.

ALGORITHM=INSTANT is the default ALTER algorithm since
MySQL 8.0.28 for operations such as add or drop column
if no ALGORITHM is explicitly defined by the user. If a
column order ALTER isrun in the same query with an add
or drop, the expectation is that InnoDB should use either
COPY or INPLACE which are the lowest common
denominator algorithms that are supported for the
aforementioned operations. However, this is not the case
currently as a drop done together with a column order change
leads to using ALGORITHM=INSTANT by default under-the-hood:

    mysql> ALTER TABLE test.`t` DROP COLUMN c3, MODIFY COLUMN c4 datetime after c1;
    Query OK, 0 rows affected (0.08 sec)
    Records: 0  Duplicates: 0  Warnings: 0

This causes a silent corruption of the table involved, that leads to
a database crash when a database restart is performed. The cause
of the corruption is a mismatch between the physical ordering
and ordinal ordering of the columns. E.g. for the case tested
in which the position of column c4 was changed in the ALTER
above, when we query the system tables we find that the
ordinal position of c4 puts it in the second position, in
between columns c1 and c2, but its physical position places it
after c2 instead of before. See below:

mysql> select table_id,ordinal_position,name,se_private_data from mysql.columns where table_id IN (select table_id  from mysql.columns where name='c1') order by table_id,ordinal_position;
+----------+------------------+---------------------------+-----------------------------------+
| table_id | ordinal_position | name                      | se_private_data                   |
+----------+------------------+---------------------------+-----------------------------------+
|      371 |                1 | c1                        | physical_pos=3;table_id=1067;     |
|      371 |                2 | c4                        | physical_pos=6;table_id=1067;     |
|      371 |                3 | c2                        | physical_pos=4;table_id=1067;     |
|      371 |                4 | DB_ROW_ID                 | physical_pos=0;table_id=1067;     |
|      371 |                5 | DB_TRX_ID                 | physical_pos=1;table_id=1067;     |
|      371 |                6 | DB_ROLL_PTR               | physical_pos=2;table_id=1067;     |
|      371 |                7 | !hidden!_dropped_v1_p5_c3 | physical_pos=5;version_dropped=1; |
+----------+------------------+---------------------------+-----------------------------------+
7 rows in set (0.03 sec)

The fix for this issue is to remove the flags for position ordering
of columns from the constant that defines the flags of operations for
which ALGORITHM=INSTANT is allowed. The fix solves the bug
reported in https://bugs.mysql.com/bug.php?id=113442 . This bug was
introduced by the commit 1cb4e7ff and affects MySQL versions
8.0.28+.

This contribution is under the OCA signed by Amazon and covering
submissions to the MySQL project.