Bug#36084930 : Firewall Stored procedures failed to update after upgrade
Problem:
After the upgrade process, the firewall
stored procedures shall be upgraded too
but because of the if condition below
in sql/dd/impl/upgrade/server.cc

if (has_old_firewall_tables && !has_new_firewall_tables)

the stored procedures are never upgraded
when upgrading from MySQL 8 to any other
8.X or 8.0.X.

Analysis:
The "if" condition mentioned above allows
the upgrade to happen only if the user
is upgrading from MySQL 5.

That if condition needs to changed such
that upgrade of procedure happens each
time server is upgraded.

Also, since MySQL 8.2, the Firewall plugin
now supports custom schema for plugin to
use. The upgrade shall update the procedures
which are present in the Firewall schema,
not just in mysql schema.

Fix:
The fix is to make sure that upgrade of
Firewall's procedures is triggered if the
plugin is already present.
And the upgrade process must upgrade the
stored procedures wherever they are
present, either in mysql schema or the
custom schema pointed by --mysql-firewall-schema
which was added in MySQL 8.2.

Change-Id: I7863ead80c2b335223e79cf1f7ab8bd59d332662