Bug#36190386: identical tables are checked in HashJoin because of no STOP_AT_MATERIALIZATION
In queries that materialize rows in a temporary table before
performing hash join or streaming aggregation, data may be copied
twice from the temporary table to the join buffer or aggregation
buffer. It does not cause any wrong results, but it makes it use the
buffer space inefficiently and it may affect performance negatively.

The bug is in WalkTablesUnderAccessPath(), which visits tables in
MATERIALIZE access paths twice. First when it sees the MATERIALIZE
access path, and then again when it visits the table_path member of
the MATERIALIZE access path.

Fixed by not visiting the table when seeing the MATERIALIZE path, and
only visit it when seeing the table_path below MATERIALIZE.

Change-Id: I6f32b851d74941ac0aa4281373dc12bd7e6af1dd