Bug#36108397 Upgrade to latest protobuf library [patches]
Followup patch:

We no longer need to support el6, so remove our own homegrown linker
scripts for the protobuf libraries. This fixes problems with ODR
violations when running protoc with gRPC plugins. It also fixes bad
symbol bindings for our plugins which link with the protobuf shared
libraries.

We switch from STATIC to SHARED abseil libraries, in order to make
linking consistent for MySQL and plugins, and for protoc and its
(gRPC) plugins.

One extension to extra/protobuf/protobuf-24.4/src/libprotobuf.map
was needed for UBSAN builds with gcc/clang.

Change-Id: Icaddbfc1dbebe26cfe1d79a7acb39d329dd66e25