Bug#36270629 Certificate expires earlier than expected
Certificate lifetimes are supposed to have a fixed number of days
plus a random amount of extra time. The random extra time is
generated using the OpenSSL function RAND_bytes() and casting the
result to an int value. Because this int was signed, it would
sometimes be negative, causing the extra time to be subtracted
rather than added.

Fix the problem by using an unsigned data type to hold the value.

Also correct a spelling error in output of ndb_sign_keys.

Change-Id: If89fe54c824085b1cbc56e54bb6a5891db71ef07