Wl#16071 JSON data type and file format inference in Lakehouse
This commit enables inference for JSON data type and JSON format on Lakehouse tables.
The inference is applicable to all JSON columns that are presented in our previously
supported formats (CSV, Parquet, Avro) and now in JSON format.
When running Autoschema on external objects, the type, name and other statistics of
each JSON column is correctly broadcasted to client.

Change-Id: I1c4141e4a391800ee316590e974ba5999a35f258