Bug#36142157 Audit Log: filtering issue for              "general_sql_command.str": "error"
Problem: The audit log filtering does not log the error
statements in the audit log file when the audit filter
specifically contains the "error" type. But if the filter
does not have this error type, the error statements can be
seen in the audit log.

Fix: The "error" type length is always set to zero before
the output is outstreamed to the audit log file; as a result
no entry for the "error" statement is detected. The fix is
to set the "error" type length to a correct value instead
of zero.

Change-Id: I5961679e1794e0f707eb861cd4f3c8d4a3276cc4