Bug#36174938 innodb doesn't call fsync on directories after unlink, rename, open(o_create)
Some file system operations performed by InnoDB on Unixish platforms
alter directory entries, and should fsync the parent directory,
before making further progress.

Without syncing directories, one can not assume that the results
of rename, unlink, and similar operations are already persisted
in the file system.

Removing log_flush_directory_low, as it was only in log_rename_file_low.
The function it was serving is now performed by the code added in this patch.
This should fix an issue related to WSL1 not supporting fdatasync:
Bug#36418662 Server fails to start on WSL1.

Change-Id: I50bd8ea67c6b0e2b56880a81491d777362a06bd1