BUG#34259810: gr_view_change_uuid_check_configuration failing on daily trunk
Test gr_view_change_uuid_check_configuration was sporadically
failing on error log content assert.
The scenario is:
1. Configure a invalid gr_view_change_uuid_check_configuration
   value.
2. Restart the server with Group Replication configured to start
   automatically, which does fail due to the invalid value.
3. Assert that there is the expected error message on the server
   error log.
The server is restarted with `restart_mysqld.inc` which will wait
for the server to be up, that is, the server is accepting client
connections. This point is before the Group Replication automatic
start, thence the error situation may not yet been reached when the
error log content is asserted.

To solve this timing issue, we only assert the error log after the
expected message is present on `performance_schema.error_log` table,
that will ensure that error situation was already reached.

Change-Id: I5a49ba82f500f2ec95d633baa4e9039161454adc