Bug#36104667: Assertion failure in Sort_param::make_sortkey with hypergraph_optimizer
When using the hypergraph optimizer, an assertion failure was
sometimes seen when sorting NULL-complemented rows from a left or
right outer join.

The problem was that Filesort did not check table->has_null_row()
before saving or restoring values of non-nullable columns. If the row
was NULL-complemented, this could lead to attempts to store garbage in
the sort buffer, and possibly writing garbage beyond the bounds of the
sort buffer.

This did not happen with the old optimizer, because it always streams
the rows through a temporary table before sorting rows after a join,
so Filesort wouldn't see a NULL-complemented row in this case, but
rather an ordinary row with NULL values in a temporary table.

Also, it did not happen when Filesort used packed addon fields, as
that code path already took NULL-complemented rows into account. It
only happened with non-packed addon fields.

Fixed by handling NULL-complemented rows the same way for non-packed
addon fields as for packed addon fields.

Change-Id: Ia2d2dc2f579546db3dd79c150f993f3f08ba652a