Bug#36257081: Assertion `!OrderItemsReferenceUnavailableTables              (path, used_tables_before_repla
For the sql syntax that is supported only with HeatWave, parser sets that
such a query could be executed only in secondary engine when it looks at
this syntax. When parsing a view definition, this setting is changed
to reflect that of view definition's setting . As a result,
if the query invoking the view has the special syntax while
the view definition itself does not, query is mistakenly marked
to be executed even in primary engine leading to problems.
Solution is to set that a query can be executed only in secondary engine
while parsing the view definition if the view itself has such a syntax.
Else, the original setting for the entire query is left as is.

Change-Id: I7103d56b6641cebbf435fbc02aeffcd4e2276874