Bug#36288425: Hypergraph: Wrong result with range queries               and multi-valued indexes
Range query on any key-part of a multi-valued index does not
work as expected because handler does not do the required checks
after retrieving the row.
E.g For a range condition like this one - id > 10 and id < 20
Handler fetches the record with id > 10. But it does not check
for id < 20 even though range optimizer has specified an end
range for the range condition.
Range queries are not supported on multi-valued indexes. However
range optimizer does construct a range tree which is not exact
when the index is used for functions like JSON_OVERLAPS/JSON_CONTAINS
etc. And care is taken to add the filter on top of the range
scan for such cases.
In the case where the range is on a keypart (of a composite
multi-valued index) which is not an array, range optimizer
constructs an exact range. However, as the handler does not
do the checks for the "end_range", we always retrieve more
rows than that are qualified by the range condition.
Wrong results are not returned in case of old optimizer
as it always places the filter on top of the range scan.
We do the same now for hypergraph optimizer too i.e if a range
tree is constructed on any keypart of a multi-valued index,
range tree is always marked as inexact.

Change-Id: I756abf242fa237ce5759ba9fc43d64ab9bb96cc6