BUG#35918034: Relax GR member join version compatibility rules on 8.4 LTS
Currently in Group Replication, a server can only join a group if
the server version is equal or higher than the lowest group member.
Example, on a group with S1 (8.2.0), S2 (8.2.0) and S3 (8.3.0),
to be allowed to join, the joining member must be 8.2.0 or higher.

With the new release model supporting LTS and innovation releases,
inplace downgrade should be supported, but only within one LTS
release series.
For server versions 8.4.0+ (8.4.1, 8.4.2, etc.) inplace downgrade
shall be allowed.

The same reasoning applies to cross-version groups, that is, a
server 8.4.0+ should be allowed to join a 8.4.1 or higher 8.4
patch version.
Example: on a group with S1 (8.4.1) and S2 (8.4.2), a server with
version 8.4.0 is allowed to join.

For this rule to apply all servers version must be 8.4, that is,
if a higher major version server is part of the group it does not
apply.
Example: on a group with S1 (8.4.1) and S2 (9.0.0), a server with
version 8.4.0 is not allowed to join.

The change of this rule will not impact the primary election
algorithm.
On a primary failover, the elected primary will be from the lowest
version.
Example:
 1) A group with S1 Primary (8.4.1) and S2 Secondary (8.4.2),
    a server with version 8.4.0 is allowed to join and will become
    S3 Secondary (8.4.0).
 2) If S1 Primary leaves the group, the new primary will be S3
    (8.4.0).

Likewise, only the lowest version servers can be used on
group_replication_set_as_primary().
Example:
 1) A group with S1 Primary (8.4.1) and S2 Secondary (8.4.2),
    a server with version 8.4.0 is allowed to join and will become
    S3 Secondary (8.4.0).
 2) group_replication_set_as_primary() will only accept S3
    (8.4.0) UUID.

Change-Id: Ifa0d64127ea23bea52702756c5309326fa551664