WL#15831: Step 1/4: Removing parser keywords
This step removes the keywords and rules pertaining the commands that
were deprecated and will now be removed in this patch.
It also does minor test cleanup dealing with reserved keywords that
will now disappear.
Minor code renames are also done to modernize the parser code in terms
of terminology.

Change-Id: Id188dfce2ae74d912c3d982f775095c0c023bb38