WL#15831: Step 2/4: Changes to client tools and command structures
This patch does minor changes to mysql client to modernize the list of
words it uses for completion.

For mysqladmin we replace the usage of RESET MASTER with the new
command RESET BINARY LOGS AND GTIDS. The tool will however use the old
command if the version of the server it is contacting does not support
it.

We also do minor renames for command enum values in the server.

Finally, we also remove the deprecated status variables that contain
the master/slave terminology.

Change-Id: If609924fc99a61e7e373f1ed01a3e758627cf66f