Bug#36292498: Compile MySQL with clang 18 [noclose]
Silence some false -Wdocumentation warnings with clang 18, such as:

storage/innobase/dict/dict0dd.cc:2729:2: warning: '@tparam' command
used in a comment that is not attached to a template declaration
[-Wdocumentation]

It is a known[1][2] issue that -Wdocumentation produces false warnings
for @tparam on templates that have explicit instantiations, and
my_compiler.h already has the macro
MY_COMPILER_CLANG_WORKAROUND_TPARAM_DOCBUG to silence it.

Use the macro two more places in dict0dd.cc. The declarations of the
affected templates already used the macro, but with clang 18 it was
also needed for the definitions.

The warnings in item_func.cc were fixed by removing the explicit
instantiation of the templates, as they were not really needed.

[1] https://bugs.llvm.org/show_bug.cgi?id=35144
[2] https://github.com/llvm/llvm-project/issues/52977

Change-Id: I4629f7195b11ac0c244ae3f55b1a55c0309f7338