Bug#36215931 The binlog_gtid_tag_and_mode test fails sporadically
Problem
-------

The 'binlog_gtid_tag_and_mode' cannot be run twice, second run fails
with an error. Moreover, the same test sporadically fails on PB2 with
'ER_CANT_SET_GTID_MODE'.

Analysis / Root-cause analysis
------------------------------

Firstly, the 'binlog_gtid_tag_and_mode' requires gtid_executed to be
cleared out before the test starts.
Secondly, spurious failures of the test case 3 is expected. Counter
that is responsible for counting the number of sessions
executing transactions with GTID_NEXT set to AUTOMATIC:tag may
be updated in the THD destructor. It happens in case session
is destroyed but has GTID_NEXT set to AUTOMATIC:tag. It is ok
to decrement counter in THD destructor. If the user wants to
change GTID_MODE, the user should wait until session is destroyed
and counter is decremented or should explicitly set GTID_NEXT
to 'AUTOMATIC'. This would force earlier update of the counter
and allow for successful change of the GTID mode.

Solution
--------

The 'binlog_gtid_tag_and_mode' requires gtid_executed to be cleared.
Added 'RESET BINARY LOGS AND GTIDS' command at the beginning of the
test. Also, we cannot ensure that the counter
responsible for counting the number of sesssions having GTID_NEXT
set to AUTOMATIC:tag, is updated before the session disconnects.
Nevertheless, we want to check that the counter is updated correctly.
Therefore, in test case 3, we wait until setting GTID mode to
OFF_PERMISSIVE succeeds.

Change-Id: I3241bb1447a54f02c113f3196d00cccb9a41bb5a