Improved the way my_plugin_log_message() handles plugin_ptr argument.
This patch, helps MySQL to be more resilient and prevent crashes when plugins misbehave and use the function "my_plugin_log_message" with a "nullptr".
Also, with this patch MySQL will not crash, instead it will log a message telling it received and unexpected nullptr and does a best-effort to log original plugin's message.