WL#15742 connection-sharing without connection-pool
Server-side connections that are sharable are currently placed into the
connection-pool which is limited by "max_idle_server_connections".

If "max_idle_server_connections" is 0, connection-sharing is disabled.

Moving a connection into the connection-pool, means that moving it out
of the pool needs to restore the earlier session-state of system-vars,
trx-state, ... (even if the connection wasn't used by another
connection).

This WL changes how sharable connections are maintained:

1. the max_idle_server_connection only controls server-side connections
   where the client side is closed
2. sharable connections don't need to be restored if they weren't used
   by another connection.

Result is that connection-sharing and rw-splitting have less overhead.

Change
======

- new REST API element: stashedServerConnections
- store sharable connections on "the stash", seperate from
  the normal connection pool for dangling server-connections.
- simplified the connection-class that's stored in the connection-pool
  into the routing_connections library
- max_idle_server_connections=0 is allowed now
- moved headers that are shared between connection-pool and routing into
  mysqlrouter/
- moved TlsSwitchable to mysqlrouter/connection_base.h

Change-Id: I2c7b280ba06c66daccd731e551d0655a49691fea