Bug#36453684 Always use same subscription settings for ndb_apply_status
When subscribing to changes in mysql.ndb_apply_status  different
settings are used depending on wheter --ndb-log-apply-status is ON or
OFF. Since the --ndb-log-apply-status variable can be changed at runtime
and the subscriptions are not recreated at that time, the different
settings don't have the desired effect when changed at runtime.

The changes to ndb_apply_status which are received from the cluster and
written to the binlog in this way are often referred to as "reflected"
ndb_apply_status changes.

Currently the ndbcluster plugin:
 - always subscribe to changes in ndb_apply_status table
 - always update all columns of ndb_apply_status when
   the applier writes to it.
 - writes the changes to ndb_apply_status to binlog
   when --ndb-log-apply-status is ON

This means that the difference between enabling --ndb-log-apply-status=1
dynamically at runtime rather than from start of the MySQL proccess is
the format used when writing the ndb_apply_status updates to the binlog.
When enabling at runtime, the writes will still be done using UPDATE
format while WRITE was intended. The missing use of FULL has no effect.

Using WRITE format makes the binlog image slightly smaller and should
thus be preferred.

To fix this inconsistency and make the behaviour more distinct, the
settings are further hardcoded to always use WRITE format. Whether or
not to use the full event does not matter and changing that would
primarily cause additional events to be created in existing clusters.
Also the cleanup of old events are improved to drop both events should
they exist - this is primarily used when cleaning up a failed attempt to
create table and events.

Change-Id: I480f11babbf4dd0334d7d199a38cc2abb2d10fb3