Unified behaviour when calling "plugin->deinit" for all PUBLIC plugin's type by passing a reference to then plugin rather than `nullptr`.
Before calling "plugin->deinit", MySQL verifies if for the given plugin's type there is a special function that handles the deinit phase.
This is achieved by verifiying the array "plugin_type_deinitialize" in `sql/sql_plugin.cc`.
Currently, there are 3 special functions: "ha_finalize_handlerton" (MYSQL_STORAGE_PLUGIN),
"finalize_schema_table" (MYSQL_INFORMATION_SCHEMA_PLUGIN) and "finalize_audit_plugin" (MYSQL_AUDIT_PLUGIN).

If there is not a special fuinction then MySQL check if the plugin exposes a "deinit" function and executes it passing the "plugin" as argument.
Otherwise, MySQL calls the "special function" passing the plugin as argument. Then the special function check if the plugin exposes a "deinit"
function and calls it but it always pass `nullptr` as argument rather than the plugin.

Because of that, plugin's deinit cannot use plugin's struct and need to add "maybe_unused" (or "__attribute__((unused))") to the signature.