Bug#36308318 GTID_NEXT accepts Invalid value and also select shows wrong value
Problem
-------

GTID_NEXT does not report an error on invalid inputs when set
to AUTOMATIC with a tag. It accepts a valid part of a string
and truncates the rest of the tag input. Also, when GTID_NEXT
is set to a correct value, SELECT statement is printing garbage
after the tag.

Analysis
--------

Parsing functions were prepared to work in several
contexts:
- tags followed by other GTIDs (GTID sets)
- tags of a standalone GTID.
Therefore, tag parsing functions do not report an error
when a correct sign is detected at the end (comma,
colon) and printing functions do not add a terminating null
character at the end of the string.

Solution
--------

Changed Gtid_specification implementation so that:
- parsing function is reporting an error in case
  we parse GTID_NEXT set to AUTOMATIC with a tag and
  extra characters appear after a valid input
- printing function is adding a null character after
  a tag when GTID_NEXT is set to AUTOMATIC with a
  tag

Change-Id: Ie5997044e5df4c00ebec86394a5bfd99be5088f7