Bug#36098954: Assertion GraphSimplifier::EdgesAreNeighboring failed with hypergraph_optimizer
A query failed when running with the hypergraph optimizer enabled and
with optimizer_max_subgraph_pairs set to a low value.

It was caused by a division by zero during graph simplification. A
similar problem was fixed in bug#34854369, but that bug only adjusted
the cardinality of base relations to be non-zero. It did not take into
account that a predicate with selectivity zero could also cause
cardinality estimates of zero.

This patch fixes the problem by ensuring that the cardinality is at
least 0.1 rows for each side of the join during graph simplification,
just to avoid division by zero. The choice of 0.1 rows as a minimum is
arbitrary.

Change-Id: I40c70d4d9b832228a8081693bc62183aaae861ae