Bug#36124625 HCS-11410 : Restore Dbsystem Failing with error "at least one redo file is missi
After calling rename(..) on POSIX platform one should fsync the
(destination) directory node to make sure change is persisted to disc.
Redo Log module was attempting fsync("data-dir/") instead of
fsync("data-dir/#innodb_redo/").
This could manifest as data loss and InnoDB not being able to start up,
if a power-outage (or OS crash) would happen right after rename, and
before the change to directory got persisted to disc on its own pace.
Similar problem could occur if a snapshot of the file system was taken
at that time.

The way this bug would manifest in practice is that when InnoDB needs a
new redo log file with number 280 (because 279th has become full of
data), what it does is the following (see log_files_produce_file()):

1. create "#ib_redo280_tmp" (using open(..) syscall)
2. resize "#ib_redo280_tmp"  (using fallocate(..) + fsync/fdatasync())
3. close("#ib_redo280_tmp")
4. open("#ib_redo280_tmp")
5. write the correct headers to "#ib_redo280_tmp" (using pwrite(..))
6. because we use a RAII Log_file_handle which notices the file was
   modified by writes, calls fsync()/fdatasync() before calling close()
   on the handle
7. rename "#ib_redo280_tmp"  to "#ib_redo280" (using rename() syscall)
8. fsync() or fdatasync() the parent directory which contains the
   renamed file (using a handle obtained from open(directory))
9. mark "#ib_redo279" as full (using open(..), pwrite(...) )
   (note the lack of fsync. We are ok with persisting it "later",
   but it shouldn't be "too early")
10. close("#ib_redo279")

The problem is that in step 8. we fsync wrong directory, so during
recovery the InnoDB would notice that the redo log file with the
largest number, #ib_redo279, already has LOG_HEADER_FLAG_FILE_FULL
indicating that it was full, which implies a new file with even larger
number, #ib_redo280, should have been created (which involves rename)
to accommodate new data, yet this newer file doesn't exist.
InnoDB woud fail with an error:

[ERROR] MY-013893 [InnoDB] Found existing redo log files, but at least
one is missing. It is unknown if recovery could reach physically
consistent state. Please consider restoring from backup or providing
--innodb-force-recovery > 0.

This patch fixes the bug, by calling fsync on the correct directory.

Change-Id: I87f4eb23c60b0d82876789069d1a6716734815ac