Bug#36314993 CUBE caused crash in Query_block::supported_correlated_scalar_subquery
I refer to the simplified repro with ROLLUP, not CUBE in the below.
There are two issues.

1.In the rollup query we have

  SELECT ( SELECT f2
           FROM (SELECT 1
                 GROUP BY ROLLUP(1)) AS dt1
           GROUP BY f1) AS a,
         MIN(f1)
  FROM t1;

In preparation for transforming the scalar subquery, we move the
aggregate MIN(f1) into a derived table, thusly:

  SELECT (SELECT derived_1_4.Name_exp_2
          FROM (SELECT 1
                GROUP BY '' WITH ROLLUP) AS dt1
          GROUP BY derived_1_4.Name_exp_1) AS a,
         derived_1_4.`MIN(f1)`
  FROM (SELECT MIN(t1.f1) AS `MIN(f1)`,
               t1.f1 AS Name_exp_1,
               t1.f2 AS Name_exp_2
        FROM t1) AS derived_1_4

As can be seen the original occurences of f1 and f2 in the scalar
subquery have been replaced with another (outer) reference:
derived_1_4.Name_exp_1 for f1 and derived_1_4.Name_exp_2 for f2. f1 is
induced as a hidden field in the select list along with visible f2,
i.e. it is scalar. We lost the hidden flag for f1 alias
derived_1_4.Name_exp_1, which lead to the first assert: it was no
longer a scalar subquery as expected. We now made sure to inherit the
hidden flag in Item_field::replace_item_field. To facilitate this we
had to make a copy of f1 in Query_block::transform_grouped_to_derived
lest we clobber its hidden status when calling add_item_to_list a few
lines further down.

2. When doing the replacement of t1's fields after moving it into the
derived table, we correctly detected that (hidden) f1 in the select
list and f1 in the group by list are outer references, cf. the logic
in Item_field::replace_item_field. However, we created two different
items for those two occurences, which caused another assert when
creating the tmp table; we now make sure we use only one field, cf. the
new member field in struct Item_field_replacement: m_outer_field.

Now, in this case, the whole exercise of moving the grouping into a
derived table turns out to be useless, as we can't transform the
scalar subquery anyway (correlated but no equality clause,
rollup). But we don't know that this early in the transform, so it is
done and should work.

The other repros no longer crash, and are included in the tests.

Change-Id: Iac3e35bdbf2f0385e1e4e33083f4e2d19060036d